+++ client/src/components/HomePageHeader.css
@@ -1,0 +1,105 @@
+:root {
+    --small-icon-width:35px;
+    --medium-icon-width:33px;
+    --large-icon-width:30px;
+    --header-bgcolor:#FCD8D4
+}
+  
+body {
+    background-color:#FCD8D4
+}
+  
+.logout-section {
+    padding: 10px 10px;
+    background-color: var(--header-bgcolor);
+    text-align: left;
+    /* display: flex;
+    justify-content: space-between; */
+}
+  
+.logout {
+    text-align:left;
+}
+.setting-section {
+    display: flex;
+    justify-content: flex-end;
+    background-color: var(--header-bgcolor);
+    text-align:right;
+}
+  
+.circular-icon {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 50px; /* adjust as needed */
+    height: 50px; /* adjust as needed */
+    border-radius: 50%;
+    /* background-color: #ccc; adjust as needed */
+}
+  
+.user-section {
+    display:flex;
+    justify-content: flex-start;
+    /* justify-items:start; */
+    background-color:var(--header-bgcolor);
+    text-align:left;
+}
+  .user-picture {
+    display:flex;
+    justify-content: center;
+    /* flex-grow: 1; */
+}
+  
+.user-location {
+    /* background-color: #FCD8D4; */
+    text-align:left;
+    font-size:30px;
+}
+
+.local-production {
+    text-align:left;
+    font-size:15px;
+    /* background-color: #FCD8D4; */
+}
+
+.settings-icon {
+    fill: darkgrey;
+}
+  
+@media (maxwidth: 320px) {
+    
+.logout {
+      border-radius: 2px;
+}
+    
+.settings-icon {
+      width: var(--small-icon-width);
+      
+  }
+} 
+  
+@media (min-width: 321px) and (max-width: 768px) {
+    
+.logout {
+      border-radius: 10px;
+    }
+    
+    .settings-icon {
+      width: var(--medium-icon-width);
+      
+    }
+    
+}
+  
+@media (min-width: 769px) {
+    
+    .logout {
+      border-radius: 15px;
+    }
+    
+    .settings-icon {
+      width: var(--large-icon-width);
+      
+    }
+}
+
